requirements.txt
beautifulsoup4==4.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
deprecation==2.1.0
et-xmlfile==1.1.0
filelock==3.9.0
html5lib==1.1
huggingface-hub==0.12.1
idna==3.4
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pikepdf==7.1.1
Pillow==9.4.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-magic-bin==0.4.14
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sapbdp-business-ocr-client==0.0.16
shapely==2.0.1
six==1.16.0
soupsieve==2.4
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
xlrd==2.0.1
